% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_footnotes}
\alias{clean_footnotes}
\title{clean_footnotes}
\usage{
clean_footnotes(wikitables, ...)
}
\arguments{
\item{...}{passes arguments to stringr::str_remove_all()}

\item{wiki_table}{A dataframe from which the footnotes names will be removed}
}
\value{
Cleaned dataframe
}
\description{
clean_footnotes
}
