% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_rows}
\alias{clean_rows}
\title{clean_rows}
\usage{
clean_rows(wiki_tables)
}
\arguments{
\item{wiki_table}{A list of dataframes for which the rows will be cleaned}
}
\value{
a list of cleaned dataframes
}
\description{
clean_rows
}
