% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_rows_single}
\alias{clean_rows_single}
\title{clean_rows_single}
\usage{
clean_rows_single(wiki_table)
}
\arguments{
\item{wiki_table}{A dataframe for which the rows will be cleaned}
}
\value{
Cleaned dataframe
}
\description{
clean_rows_single
}
