% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_wiki_names}
\alias{clean_wiki_names}
\title{clean_wiki_names}
\usage{
clean_wiki_names(wiki_tables, ...)
}
\arguments{
\item{...}{passes arguments to janitor::clean_names()}

\item{wiki_table}{A dataframe for which the column names will be cleaned}
}
\value{
Cleaned dataframe
}
\description{
clean_wiki_names
}
