% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_wiki_names_single}
\alias{clean_wiki_names_single}
\title{clean_wiki_names_single}
\usage{
clean_wiki_names_single(wiki_table, ...)
}
\arguments{
\item{wiki_table}{A dataframe for which the column names will be cleaned}

\item{...}{passes arguments to janitor::clean_names()}
}
\value{
a cleaned dataframe
}
\description{
clean_wiki_names_single
}
