% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_wikitable}
\alias{clean_wikitable}
\title{clean_wikitable}
\usage{
clean_wikitable(wiki_tables, ...)
}
\arguments{
\item{...}{passes arguments to read_wikitables()}

\item{table_num}{the number of the single tbl you want to clean}
}
\value{
A single tbl
}
\description{
clean_wikitable
}
\examples{

}
