% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{convert_types_single}
\alias{convert_types_single}
\title{convert_types_single}
\usage{
convert_types_single(wiki_table)
}
\arguments{
\item{wiki_table}{A dataframe for which the rows will be cleaned}
}
\value{
Cleaned dataframe
}
\description{
convert_types_single
}
