% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{empty_to_na}
\alias{empty_to_na}
\alias{special_to_na}
\title{empty_to_na}
\usage{
empty_to_na(wiki_table)

special_to_na(wiki_table)
}
\arguments{
\item{wiki_table}{A dataframe}
}
\value{
A dataframe
}
\description{
Convert empty strings and special characters to NA
}
\examples{
empty_to_na(data.frame(x = ""))
special_to_na(data.frame(x = c("?", "le#t"), y = c("hi", "8\%")))
}
