% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_no_clean}
\alias{read_no_clean}
\title{read_no_clean}
\usage{
read_no_clean(url, table_number = 1)
}
\arguments{
\item{url}{A character vector of the url of a wikipedia page containing a table. Default is 1.}

\item{table_number}{A numeric denoting which number table the desired table is on the page (e.g. the second table on the page)}
}
\value{
Dataframe of wikipedia table
}
\description{
read_no_clean
}
\examples{
read_no_clean('https://en.wikipedia.org/wiki/List_of_most-followed_Instagram_accounts')
read_no_clean('https://en.wikipedia.org/wiki/List_of_cryptids', 3)
read_no_clean('https://en.wikipedia.org/wiki/List_of_Pixar_films')
}
