% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table_raw}
\alias{read_table_raw}
\title{read_table_raw}
\usage{
read_table_raw(url, table_number = 1, replace_linebreak = ", ")
}
\arguments{
\item{url}{A character vector of the url of a wikipedia page containing a table. Default is 1.}

\item{table_number}{A numeric denoting which number table the desired table is on the page (e.g. the second table on the page)}

\item{replace_linebreak}{A character string denoting what linebreaks within dataframe cells are to be replaced with. Default is ' ,'.}
}
\value{
Dataframe of wikipedia table
}
\description{
read_table_raw
}
\examples{
read_table_raw('https://en.wikipedia.org/wiki/List_of_most-followed_Instagram_accounts')
read_table_raw('https://en.wikipedia.org/wiki/List_of_cryptids', 3)
read_table_raw('https://en.wikipedia.org/wiki/List_of_Pixar_films')
}
