% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_wikinodes}
\alias{read_wikinodes}
\title{read_wikinodes}
\usage{
read_wikinodes(url, replace_linebreak = ", ", ...)
}
\arguments{
\item{url}{A character vector of the url of a wikipedia page containing a table. Default is 1.}

\item{replace_linebreak}{A character string denoting what linebreaks within dataframe cells are to be replaced with. Default is ' ,'.}

\item{...}{passes arguments to read_wikitables()}
}
\value{
list of xlm nodes
}
\description{
read_wikinodes
}
\examples{

}
