% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_wikitables}
\alias{read_wikitables}
\title{read_wikitables}
\usage{
read_wikitables(url, ...)
}
\arguments{
\item{url}{A character vector of the url of a wikipedia page containing a table. Default is 1.}

\item{...}{passes arguments from read_wikinodes}
}
\value{
list of tbls
}
\description{
read_wikitables
}
\examples{

}
