% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{remove_footnotes}
\alias{remove_footnotes}
\title{remove_footnotes}
\usage{
remove_footnotes(wiki_table, ...)
}
\arguments{
\item{wiki_table}{a dataframe}

\item{...}{arguments passed to \code{\link{add_na}}}
}
\value{
a cleaned dataframe
}
\description{
remove_footnotes
}
