% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{remove_footnotes_single}
\alias{remove_footnotes_single}
\title{remove_footnotes_single}
\usage{
remove_footnotes_single(wiki_table, ...)
}
\arguments{
\item{wiki_table}{a dataframe}

\item{...}{Passes arguments to add_na()}
}
\value{
a cleaned dataframe
}
\description{
remove_footnotes_single
}
