% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{add_na}
\alias{add_na}
\alias{add_na_single}
\alias{special_to_na_single}
\alias{special_to_na}
\title{add_na_single}
\usage{
add_na_single(wiki_table, to_na = "", special_to_na = TRUE)

add_na(wiki_tables, to_na = "", special_to_na = TRUE)

special_to_na_single(wiki_table)

special_to_na(wiki_tables)
}
\arguments{
\item{wiki_table}{a dataframe}

\item{to_na}{A character string that when solitary in a dataframe cell is to be converted to NA. Default is "".}

\item{special_to_na}{A boolean denoting whether solitary special characters in dataframe cells are to be converted to NA. Default is TRUE.}

\item{wiki_tables}{a list of dataframes}
}
\value{
Cleaned dataframe

a list of cleaned dataframes

a cleaned dataframe

a list of cleaned dataframes
}
\description{
add_na_single
}
