% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_wiki_names}
\alias{clean_wiki_names}
\title{clean_wiki_names}
\usage{
clean_wiki_names(wiki_tables, ...)
}
\arguments{
\item{wiki_tables}{a list of dataframes for which the column names will be cleaned}

\item{...}{passes arguments to janitor::clean_names()}
}
\value{
a list oc cleaned dataframes
}
\description{
clean_wiki_names
}
