% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_table.R
\name{remove_footnotes}
\alias{remove_footnotes}
\title{remove_footnotes}
\usage{
remove_footnotes(wiki_tables, ...)
}
\arguments{
\item{...}{Passes arguments to add_na()}

\item{wiki_table}{A dataframe}
}
\value{
Cleaned dataframe
}
\description{
remove_footnotes
}
